@@ -9,49 +9,26 @@ plugins {
 }
 
 repositories {
-    mavenLocal()
-    maven {
-        url = 'https://maven.vaadin.com/vaadin-prereleases'
-    }
-
+    mavenCentral()
     maven {
         url = 'https://maven.vaadin.com/vaadin-addons'
-    }
-
-    maven {
-        url = 'https://oss.sonatype.org/content/repositories/vaadin-snapshots/'
-    }
-
-    maven {
-        url = 'https://repo.maven.apache.org/maven2'
     }
 }
 
 dependencies {
-    implementation 'com.vaadin:vaadin-compatibility-server:8.0.0.alpha2'
-    implementation 'com.vaadin:vaadin-compatibility-shared:8.0.0.alpha2'
-    implementation 'com.vaadin:vaadin-compatibility-client-compiled:8.0.0.alpha2'
-    implementation 'com.vaadin:vaadin-themes:8.0.0.alpha2'
-    implementation 'javax.servlet:javax.servlet-api:3.0.1'
-    implementation 'commons-beanutils:commons-beanutils:1.9.2'
-    implementation 'log4j:log4j:1.2.9'
-    implementation 'org.slf4j:slf4j-api:1.7.7'
-    implementation 'org.slf4j:slf4j-simple:1.7.7'
-    implementation 'org.slf4j:slf4j-log4j12:1.7.7'
-    implementation 'commons-httpclient:commons-httpclient:3.1'
-    implementation 'org.apache.commons:commons-lang3:3.1'
-    implementation 'org.json:json:20140107'
-    implementation 'org.codehaus.jackson:jackson-jaxrs:1.9.4'
-    implementation 'com.sun.xml.security:xml-security-impl:1.0'
-    testImplementation 'junit:junit:4.7'
+    implementation 'com.vaadin:vaadin:24.1.0'
+    implementation 'jakarta.servlet:jakarta.servlet-api:6.0.0'
+    implementation 'org.apache.logging.log4j:log4j-api:2.20.0'
+    implementation 'org.apache.logging.log4j:log4j-core:2.20.0'
+    implementation 'org.apache.commons:commons-lang3:3.12.0'
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.10.0'
 }
 
 group = 'com.devopsdemo.tutorial'
-// version = '2.0'
-// description = 'Vaadin Sampleapp example'
-sourceCompatibility = '1.8'
+sourceCompatibility = '21'
 
-configurations.all {
+tasks.withType(JavaCompile) {
+    options.encoding = 'UTF-8'
 }
 
 publishing {
@@ -60,8 +37,4 @@ publishing {
             from(components.java)
         }
     }
-}
-
-tasks.withType(JavaCompile) {
-    options.encoding = 'UTF-8'
 }
