@@ -1,26 +1,21 @@
 package com.devopsdemo.tutorial.addressbook.backend;
 
-import org.apache.commons.beanutils.BeanUtils;
-
+import java.time.LocalDate;
 import java.util.*;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-/** Separate Java service class.
+/**
  * Backend implementation for the address book application, with "detached entities"
- * simulating real world DAO. Typically these something that the Java EE
- * or Spring backend services provide.
+ * simulating real-world DAO. Typically, these are provided by Java EE or Spring backend services.
  */
-// Backend service class. This is just a typical Java backend implementation
-// class and nothing Vaadin specific.
 public class ContactService {
 
-    // Create dummy data by randomly combining first and last names
-    static String[] fnames = { "Peter", "Alice", "John", "Mike", "Olivia",
+    private static final String[] fnames = { "Peter", "Alice", "John", "Mike", "Olivia",
             "Nina", "Alex", "Rita", "Dan", "Umberto", "Henrik", "Rene", "Lisa",
             "Linda", "Timothy", "Daniel", "Brian", "George", "Scott",
             "Jennifer" };
-    static String[] lnames = { "Smith", "Johnson", "Williams", "Jones",
+    private static final String[] lnames = { "Smith", "Johnson", "Williams", "Jones",
             "Brown", "Davis", "Miller", "Wilson", "Moore", "Taylor",
             "Anderson", "Thomas", "Jackson", "White", "Harris", "Martin",
             "Thompson", "Young", "King", "Robinson" };
@@ -29,55 +24,39 @@ public class ContactService {
 
     public static ContactService createDemoService() {
         if (instance == null) {
-
             final ContactService contactService = new ContactService();
 
-            Random r = new Random(0);
-            Calendar cal = Calendar.getInstance();
+            var r = new Random(0);
             for (int i = 0; i < 100; i++) {
-                Contact contact = new Contact();
+                var contact = new Contact();
                 contact.setFirstName(fnames[r.nextInt(fnames.length)]);
                 contact.setLastName(lnames[r.nextInt(fnames.length)]);
                 contact.setEmail(contact.getFirstName().toLowerCase() + "@"
                         + contact.getLastName().toLowerCase() + ".com");
                 contact.setPhone("+ 358 555 " + (100 + r.nextInt(900)));
-                cal.set(1930 + r.nextInt(70),
-                        r.nextInt(11), r.nextInt(28));
-                contact.setBirthDate(cal.getTime());
+                contact.setBirthDate(LocalDate.of(1930 + r.nextInt(70),
+                        r.nextInt(11) + 1, r.nextInt(28) + 1));
                 contactService.save(contact);
             }
             instance = contactService;
         }
-
         return instance;
     }
 
-    private HashMap<Long, Contact> contacts = new HashMap<>();
+    private final Map<Long, Contact> contacts = new HashMap<>();
     private long nextId = 0;
 
     public synchronized List<Contact> findAll(String stringFilter) {
-        ArrayList arrayList = new ArrayList();
-        for (Contact contact : contacts.values()) {
-            try {
-                boolean passesFilter = (stringFilter == null || stringFilter.isEmpty())
-                        || contact.toString().toLowerCase()
-                                .contains(stringFilter.toLowerCase());
-                if (passesFilter) {
-                    arrayList.add(contact.clone());
-                }
-            } catch (CloneNotSupportedException ex) {
-                Logger.getLogger(ContactService.class.getName()).log(
-                        Level.SEVERE, null, ex);
+        var filteredContacts = new ArrayList<Contact>();
+        for (var contact : contacts.values()) {
+            boolean passesFilter = (stringFilter == null || stringFilter.isEmpty())
+                    || contact.toString().toLowerCase().contains(stringFilter.toLowerCase());
+            if (passesFilter) {
+                filteredContacts.add(contact);
             }
         }
-        Collections.sort(arrayList, new Comparator<Contact>() {
-
-            @Override
-            public int compare(Contact o1, Contact o2) {
-                return (int) (o2.getId() - o1.getId());
-            }
-        });
-        return arrayList;
+        filteredContacts.sort(Comparator.comparingLong(Contact::getId).reversed());
+        return filteredContacts;
     }
 
     public synchronized long count() {
@@ -92,12 +71,6 @@ public class ContactService {
         if (entry.getId() == null) {
             entry.setId(nextId++);
         }
-        try {
-            entry = (Contact) BeanUtils.cloneBean(entry);
-        } catch (Exception ex) {
-            throw new RuntimeException(ex);
-        }
         contacts.put(entry.getId(), entry);
     }
-
 }
