@@ -1,107 +1,87 @@
 package com.devopsdemo.utilities;
 
 import java.io.InputStream;
-import java.util.Enumeration;
 import java.util.Locale;
 import java.util.Properties;
 import java.util.ResourceBundle;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-/***
- * Looks up a resource named 'name' in the classpath. The resource must map
- * to a file with .properties extention. The name is assumed to be absolute
- * and can use either "/" or "." for package segment separation with an
- * optional leading "/" and optional ".properties" suffix. Thus, the
- * following names refer to the same resource:
- *
- * @author Seshagiri Sriram
- * @version 1.0 
- * @param name classpath resource name [may not be null] 
- * @param loader classloader through which to load the resource [null 
- * is equivalent to the application loader]
- * @return resource converted to java.util.Properties [may be null if the 
- * resource was not found and THROW_ON_LOAD_FAILURE is false] 
- * @throws IllegalArgumentException if the resource was not found and
- * THROW_ON_LOAD_FAILURE is true 
+
+/**
+ * Utility class for loading properties from files or resource bundles.
  */
+public class PropertyLoader {
 
-public class PropertyLoader 
-{
+    private static final boolean THROW_ON_LOAD_FAILURE = true;
+    private static final boolean LOAD_AS_RESOURCE_BUNDLE = false;
+    private static final String SUFFIX = ".properties";
 
-	private static final boolean THROW_ON_LOAD_FAILURE = true; 
-	private static final boolean LOAD_AS_RESOURCE_BUNDLE = false; 
-	private static final String SUFFIX = ".properties";
+    /**
+     * Logger enabled for the current class
+     */
+    private static final Logger LOG = LoggerFactory.getLogger(PropertyLoader.class);
 
-	/**
-	 * Logger enabled for the current class
-	 */
-	private static final Logger LOG =LoggerFactory.getLogger(PropertyLoader.class);
+    /**
+     * Loads properties using the current thread's context classloader.
+     *
+     * @param name the name of the properties file
+     * @return the loaded properties
+     */
+    public static Properties loadProperties(final String name) {
+        return loadProperties(name, Thread.currentThread().getContextClassLoader());
+    }
 
-	/** 
-	 * A convenience overload of {@link #loadProperties(String, ClassLoader)} 
-	 * that uses the current thread's context classloader. 
-	 */
+    /**
+     * Loads properties from a file or resource bundle.
+     *
+     * @param names  the name of the properties file
+     * @param loader the classloader to use
+     * @return the loaded properties
+     */
+    public static Properties loadProperties(String names, ClassLoader loader) {
+        if (names == null) {
+            throw new IllegalArgumentException("null input: name");
+        }
 
-	public static Properties loadProperties (final String name) { 
-		return loadProperties (name, Thread.currentThread ().getContextClassLoader ()); 
-	}
+        if (loader == null) {
+            loader = Thread.currentThread().getContextClassLoader();
+        }
 
-	@SuppressWarnings("rawtypes")
-	public static Properties loadProperties (String names, ClassLoader loader) 
-	{ 
-		String name = null;
-		ClassLoader loaders;
-		if (names == null) throw new IllegalArgumentException ("null input: name"); 
-		if (names.startsWith ("/")) name = names.substring (1); 
-		if (names.endsWith (SUFFIX)) name = names.substring (0, names.length () - SUFFIX.length ()); 
-		Properties result = null; 
-		InputStream in = null; 
-		try { 
-			if (loader == null)
-				loaders = ClassLoader.getSystemClassLoader (); 
-			if (LOAD_AS_RESOURCE_BUNDLE) 
-			{ 
-				name = name.replace ('/', '.'); // Throws MissingResourceException on lookup failures: 
-				final ResourceBundle rb = ResourceBundle.getBundle (name, Locale.getDefault (), loaders); 
-				result = new Properties (); 
-				for (Enumeration keys = rb.getKeys (); keys.hasMoreElements ();) 
-				{ 
-					final String key = (String) keys.nextElement (); 
-					final String value = rb.getString (key); 
-					result.put (key, value); 
-				} 
-			} 
-			else { 
-				name = name.replace ('.', '/'); 
-				if (! name.endsWith (SUFFIX)) name = name.concat (SUFFIX); // Returns null on lookup failures: 
-				if(loader!=null)
-					in = loader.getResourceAsStream (name); 
-				if (in != null) { 
-					result = new Properties (); 
-					result.load (in); // Can throw IOException 
-				} 
-			} 
-		} 
-		catch (Exception e) { 
-			result = new Properties ();
-			LoggerStackTraceUtil.printErrorMessage(e);
-		} 
-		finally { 
-			if (in != null) 
-				try 
-			{ 
-					in.close (); 
-			} 
-			catch (Throwable ignore) {} 
-		} 
-		if (THROW_ON_LOAD_FAILURE && (result == null)) 
-		{ 
-			// LOG exception... Do not re-throw this.. as I do not expect users to catch this exception :-) 
-			//throw new IllegalArgumentException ("could not load [" + name + "]"+ " as " + (LOAD_AS_RESOURCE_BUNDLE ? "a resource bundle" : "a classloader resource"));
-			result = new Properties ();
-		} 
-		return result;  
-	}
+        Properties result = new Properties();
+        boolean loaded = false;
 
+        try {
+            String resourcePath = names.endsWith(SUFFIX) ? names : names + SUFFIX;
+            try (InputStream in = loader.getResourceAsStream(resourcePath)) {
+                if (in != null) {
+                    // Don't process keys when loading
+                    result.load(in);
+                    loaded = true;
+                }
+            }
+
+            // If still not loaded and LOAD_AS_RESOURCE_BUNDLE is true, try as a resource bundle
+            if (!loaded && LOAD_AS_RESOURCE_BUNDLE) {
+                String bundleName = names.endsWith(SUFFIX) ? 
+                    names.substring(0, names.length() - SUFFIX.length()) : names;
+                bundleName = bundleName.replace('/', '.');
+                try {
+                    ResourceBundle rb = ResourceBundle.getBundle(bundleName, Locale.getDefault(), loader);
+                    rb.keySet().forEach(key -> result.put(key, rb.getString(key)));
+                    loaded = true;
+                } catch (java.util.MissingResourceException e) {
+                    LOG.debug("Resource bundle not found: {}", bundleName);
+                }
+            }
+        } catch (Exception e) {
+            LOG.debug("Error loading properties: {}", names, e);
+        }
+
+        if (!loaded) {
+            throw new IllegalArgumentException("Could not load properties: " + names);
+        }
+
+        return result;
+    }
 }
